# Use a faster version of goyacc that uses an interface instead of structs