… go.sum
… go/vt/sqlparser/parse_test.go
… go/vt/sqlparser/sql.go