Overview of the enhancement

Sqlparser uses goyacc to generate the sql.go file from sql.y. Goyacc currently uses structs to represent a union since go does not support any sum types yet.

Goyacc is augmented to use interfaces instead of structs for the union. The updated yacc.go is at https://github.com/GuptaManan100/tools/blob/master/cmd/goyacc/yacc.go. 

New Implementation 
BenchmarkParse1-8          83306             12198 ns/op            3626 B/op         108 allocs/op
BenchmarkParse2-8          27206             41913 ns/op           13519 B/op        390 allocs/op
BenchmarkParse2Parallel-8         110056              9495 ns/op           10757 B/op        299 allocs/op
BenchmarkParse3-8            216           5341429 ns/op         6336266 B/op        428 allocs/op
BenchmarkParseBigQuery-8              55          20029806 ns/op         4317291 B/op     192740 allocs/op

Old Implementation
BenchmarkParse1-8          78393             14656 ns/op            2447 B/op         76 allocs/op
BenchmarkParse2-8          25598             46418 ns/op            8680 B/op        266 allocs/op
BenchmarkParse2Parallel-8         128994              9346 ns/op            5913 B/op        175 allocs/op
BenchmarkParse3-8            232           5133668 ns/op         6342919 B/op        360 allocs/op
BenchmarkParseBigQuery-8              52          22312829 ns/op         2541029 B/op     133468 allocs/op

The cpu and memory profiles for all the benchmarks can be seen at https://drive.google.com/drive/folders/1ce9ZBh2_Z7-xyI2deAvrCpKQYZ9bZAya?usp=sharing